# backend/tests/contract/snapshots/api-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "BalloonCollectionResponse": {
        "additionalProperties": false,
        "properties": {
          "balloons": {
            "items": {
              "$ref": "#/components/schemas/BalloonResponse"
            },
            "title": "Balloons",
            "type": "array"
          }
        },
        "required": [
          "balloons"
        ],
        "title": "BalloonCollectionResponse",
        "type": "object"
      },
      "BalloonResponse": {
        "additionalProperties": false,
        "properties": {
          "anchor_bbox_pdf": {
            "maxItems": 4,
            "minItems": 4,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Anchor Bbox Pdf",
            "type": "array"
          },
          "center_pdf": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Center Pdf",
            "type": "array"
          },
          "collision_flags": {
            "items": {
              "enum": [
                "circle_overlap",
                "glyph_overlap",
                "glyph_circle_overlap",
                "owner_glyph_outside_circle",
                "outside_cropbox",
                "protected_overlap",
                "source_text_overlap",
                "unreadable_number",
                "invalid_leader",
                "forbidden_overlap"
              ],
              "type": "string"
            },
            "title": "Collision Flags",
            "type": "array"
          },
          "formal_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formal Number"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "inspection_item_id": {
            "title": "Inspection Item Id",
            "type": "string"
          },
          "leader_target_pdf": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Leader Target Pdf",
            "type": "array"
          },
          "page_index": {
            "title": "Page Index",
            "type": "integer"
          },
          "placement_status": {
            "enum": [
              "placed",
              "manual_required"
            ],
            "title": "Placement Status",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "sort_order": {
            "title": "Sort Order",
            "type": "integer"
          },
          "source_location_id": {
            "title": "Source Location Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "deleted"
            ],
            "title": "Status",
            "type": "string"
          },
          "suggested_number": {
            "title": "Suggested Number",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "project_id",
          "inspection_item_id",
          "source_location_id",
          "page_index",
          "suggested_number",
          "formal_number",
          "sort_order",
          "anchor_bbox_pdf",
          "leader_target_pdf",
          "center_pdf",
          "placement_status",
          "collision_flags",
          "status",
          "version"
        ],
        "title": "BalloonResponse",
        "type": "object"
      },
      "Body_QI-API-PRJ-001": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_QI-API-PRJ-001",
        "type": "object"
      },
...
      "DeleteBalloon": {
        "additionalProperties": false,
        "properties": {
          "balloon_id": {
            "format": "uuid",
            "title": "Balloon Id",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "type": {
            "const": "delete",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "balloon_id",
          "expected_version"
        ],
        "title": "DeleteBalloon",
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "blockers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blockers"
          },
          "cause_category": {
            "title": "Cause Category",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "location_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Ref"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase"
          },
          "project_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "retryable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retryable"
          },
          "severity": {
            "enum": [
              "fatal",
              "blocking",
              "review_required",
              "warning",
              "informational"
            ],
            "title": "Severity",
            "type": "string"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "workbench_ready": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workbench Ready"
          }
        },
        "required": [
          "code",
          "message",
          "severity",
          "stage",
          "location_ref",
          "cause_category"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "GenerateBalloonsRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "GenerateBalloonsRequest",
        "type": "object"
      },
...
      "MoveBalloon": {
        "additionalProperties": false,
        "properties": {
          "balloon_id": {
            "format": "uuid",
            "title": "Balloon Id",
            "type": "string"
          },
          "center_pdf": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Center Pdf",
            "type": "array"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "type": {
            "const": "move",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "balloon_id",
          "expected_version",
          "center_pdf"
        ],
        "title": "MoveBalloon",
        "type": "object"
      },
...
      "ProjectListItemResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "file_name": {
            "title": "File Name",
            "type": "string"
          },
          "last_opened_at": {
            "format": "date-time",
            "title": "Last Opened At",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "file_name",
          "created_at",
          "last_opened_at"
        ],
        "title": "ProjectListItemResponse",
        "type": "object"
      },
      "ProjectListResponse": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectListItemResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "ProjectListResponse",
        "type": "object"
      },
...
      "ProjectStatusResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectError"
              },
              {
                "type": "null"
              }
            ]
          },
          "phase": {
            "$ref": "#/components/schemas/ProjectPhase"
          },
          "project_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProcessingStage"
              },
              {
                "type": "null"
              }
            ]
          },
          "workbench_ready": {
            "title": "Workbench Ready",
            "type": "boolean"
          }
        },
        "required": [
          "phase",
          "workbench_ready",
          "retryable",
          "error"
        ],
        "title": "ProjectStatusResponse",
        "type": "object"
      },
...
      "RebuildBalloon": {
        "additionalProperties": false,
        "properties": {
          "balloon_id": {
            "format": "uuid",
            "title": "Balloon Id",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "type": {
            "const": "rebuild",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "balloon_id",
          "expected_version"
        ],
        "title": "RebuildBalloon",
        "type": "object"
      },
...
      "RenumberBalloons": {
        "additionalProperties": false,
        "properties": {
          "expected_versions": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "format": "uuid"
            },
            "title": "Expected Versions",
            "type": "object"
          },
          "ordered_balloon_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "minItems": 1,
            "title": "Ordered Balloon Ids",
            "type": "array"
          },
          "type": {
            "const": "renumber",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "ordered_balloon_ids",
          "expected_versions"
        ],
        "title": "RenumberBalloons",
        "type": "object"
      },
      "ReorderBalloon": {
        "additionalProperties": false,
        "properties": {
          "balloon_id": {
            "format": "uuid",
            "title": "Balloon Id",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "sort_order": {
            "minimum": 0.0,
            "title": "Sort Order",
            "type": "integer"
          },
          "type": {
            "const": "reorder",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "balloon_id",
          "expected_version",
          "sort_order"
        ],
        "title": "ReorderBalloon",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Quality Inspection",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "operationId": "QI-API-PRJ-006",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected internal failure.",
            "x-stable-error-codes": [
              "project_list_failed",
              "internal_server_error"
            ]
          }
        },
        "summary": "List Projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "QI-API-PRJ-001",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_QI-API-PRJ-001"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request failed transport or business validation.",
            "x-stable-error-codes": [
              "invalid_pdf",
              "request_validation_failed"
            ]
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected internal failure.",
            "x-stable-error-codes": [
              "project_intake_failed",
              "internal_server_error"
            ]
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Required dispatch or service is unavailable.",
            "x-stable-error-codes": [
              "project_dispatch_failed"
            ]
          }
        },
        "summary": "Create Project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/balloons": {
      "get": {
        "operationId": "QI-API-BAL-001",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalloonCollectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Requested resource was not found.",
            "x-stable-error-codes": [
              "project_not_found"
            ]
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request failed transport or business validation.",
            "x-stable-error-codes": [
              "request_validation_failed"
            ]
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected internal failure.",
            "x-stable-error-codes": [
              "internal_server_error"
            ]
          }
        },
        "summary": "List Balloons",
        "tags": [
          "balloons"
        ]
      }
    },
    "/api/v1/projects/{project_id}/balloons/commands": {
      "post": {
        "operationId": "QI-API-BAL-003",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-QI-Operator",
            "required": true,
            "schema": {
              "title": "X-Qi-Operator",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "delete": "#/components/schemas/DeleteBalloon",
                    "move": "#/components/schemas/MoveBalloon",
                    "rebuild": "#/components/schemas/RebuildBalloon",
                    "renumber": "#/components/schemas/RenumberBalloons",
                    "reorder": "#/components/schemas/ReorderBalloon"
                  },
                  "propertyName": "type"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/MoveBalloon"
                  },
                  {
                    "$ref": "#/components/schemas/DeleteBalloon"
                  },
                  {
                    "$ref": "#/components/schemas/RebuildBalloon"
                  },
                  {
                    "$ref": "#/components/schemas/ReorderBalloon"
                  },
                  {
                    "$ref": "#/components/schemas/RenumberBalloons"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/BalloonResponse"
                    },
                    {
                      "$ref": "#/components/schemas/BalloonCollectionResponse"
                    }
                  ],
                  "title": "Response Qi-Api-Bal-003"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Requested resource was not found.",
            "x-stable-error-codes": [
              "project_not_found",
              "balloon_not_found"
            ]
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request conflicts with current aggregate or capability state.",
            "x-stable-error-codes": [
              "balloon_version_conflict",
              "review_lock_conflict",
              "balloon_source_unavailable",
              "balloon_order_conflict",
              "review_already_confirmed"
            ]
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request failed transport or business validation.",
            "x-stable-error-codes": [
              "balloon_command_invalid",
              "request_validation_failed"
            ]
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected internal failure.",
            "x-stable-error-codes": [
              "internal_server_error"
            ]
          }
        },
        "summary": "Apply Balloon Command",
        "tags": [
          "balloons"
        ]
      }
    },
    "/api/v1/projects/{project_id}/balloons/generate": {
      "post": {
        "operationId": "QI-API-BAL-002",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-QI-Operator",
            "required": true,
            "schema": {
              "title": "X-Qi-Operator",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateBalloonsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalloonCollectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Requested resource was not found.",
            "x-stable-error-codes": [
              "project_not_found",
              "review_working_copy_not_found"
            ]
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request conflicts with current aggregate or capability state.",
            "x-stable-error-codes": [
              "review_item_set_not_frozen",
              "balloon_version_conflict",
              "review_lock_conflict",
              "balloon_source_unavailable",
              "balloon_order_conflict",
              "review_already_confirmed"
            ]
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request failed transport or business validation.",
            "x-stable-error-codes": [
              "balloon_command_invalid",
              "request_validation_failed"
            ]
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected internal failure.",
            "x-stable-error-codes": [
              "internal_server_error"
            ]
          }
        },
        "summary": "Generate Balloons",
        "tags": [
          "balloons"
        ]
      }
    },
...
  }
}